pyproject.toml
[project]
name = "rob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "robyn>=0.76.0",
]
